[day-19] boost badouralix
This implements a bunch of optimizations based on assumptions that seem
to hold. They allow us to try to match a word against a rule rather than
list all accepted words against the given rule.

In particular in part-1, we should iterate over all subrules instead of
stopping at the first matching subrule. Note that this breaks the cache
usage.

In part-2, we would need to iterate over all possible values of m and n.
In practice, that means we would superseed the match method with a
match-one-or-many-times method ( think of this as an automaton that
loops on itself ), and ensure that m > n after finding the matches.